+++ libdjvu/DjVuText.cpp
@@ -344,8 +344,8 @@ DjVuTXT::decode(const GP<ByteStream> &gbs)
   textUTF8.empty();
   int textsize = bs.read24();
   char *buffer = textUTF8.getbuf(textsize);
-  int readsize = bs.read(buffer,textsize);
-  if (readsize < textsize || textsize <= 0)
+  size_t readsize = bs.read(buffer,textsize);
+  if (readsize < (size_t)textsize || textsize <= 0)
     G_THROW( ERR_MSG("DjVuText.corrupt_chunk") );
   buffer[readsize] = 0;
   // Try reading zones
@@ -967,4 +967,3 @@ DjVuText::get_xmlText(const int height) const
 using namespace DJVU;
 # endif
 #endif
-
